Finished wall placing algorithm
To place a wall use the coordinates of two directly-diagonal board spaces as the first two parameters for the place_wall function. These coordinates define the 2x2 grid of cells that the wall will be placed between.

The final parameter of the place_wall function is the direction of the wall that will be placed. 0 is for a horizontal wall, 1 is for a vertical wall.